+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 2.0.B1
 
 * #6 Upgrade AWS libs to 1.11.913
+* #7 Integrate AWS EndpointConfiguration in the AWS config
 
 ## 1.0.RC1